Update download.py
too many changes in 1 commit I know: this is a radical change. I use GDAL to merge files instead of xml. This allows me to avoid duplicates by creating a geopackage with a unique index on column gml_id. This effectively causes an error when inserting the same gml_id. By ignoring these errors it continues to merge features that are not duplicates. Many other changes are minor. WFS 1.1.0 and 1.0.0 use numberOfFeatures instead of numberReturned/numberMatched. Also, in WFS2.0.0 when working with WGS84 (EPSG 4326) there's an inversion of axis that is now handled properly (I hope).
A few new configuration keys: version (for wfs), interval (to prevent service denial, imposes a pause between requests), uniqueid_field (to prevent duplicates).